+++ libs/arphase/ui/aps-u-ui/src/lib/directives/auto-select/auto-select.module.ts
@@ -1,0 +1,11 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+
+import { ApsAutoSelectDirective } from './auto-select.directive';
+
+@NgModule({
+  declarations: [ApsAutoSelectDirective],
+  imports: [CommonModule],
+  exports: [ApsAutoSelectDirective],
+})
+export class ApsAutoSelectModule {}