+++ libs/arphase/ui/aps-u-ui/src/lib/directives/auto-select/index.ts
@@ -1,0 +1,2 @@
+export * from './auto-select.directive';
+export * from './auto-select.module';