+++ libs/arphase/ui/aps-u-ui/src/lib/directives/index.ts
@@ -1,1 +1,2 @@
 export * from './auto-error';
+export * from './auto-select';