+++ libs/ui/u-guarantees/src/lib/components/guarantee-list/guarantee-list.component.ts
@@ -1,8 +1,16 @@
-import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
-import { Guarantee, guaranteeDateTypeOptions, GuaranteeStatus, Select } from '@ivt/c-data';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  EventEmitter,
+  Input,
+  OnChanges,
+  Output,
+  SimpleChanges,
+} from '@angular/core';
+import { Guarantee, guaranteeDateTypeOptions, GuaranteeStatus, Select, statusLabels } from '@ivt/c-data';
 import { IvtListComponent } from '@ivt/u-ui';
 
-import { menuOptions, statusOptions } from './guarantee-list.constants';
+import { colorMaps, columns, iconMaps, menuOptions, statusOptions } from './guarantee-list.constants';
 
 @Component({
   selector: 'ivt-guarantee-list',
@@ -10,7 +18,7 @@ import { menuOptions, statusOptions } from './guarantee-list.constants';
   styleUrls: ['./guarantee-list.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class GuaranteeListComponent extends IvtListComponent<Guarantee> {
+export class GuaranteeListComponent extends IvtListComponent<Guarantee> implements OnChanges {
   @Input() clearSelected: boolean;
   @Input() groupOptions: Select[] = [];
   @Input() companyOptions: Select[] = [];
@@ -21,6 +29,12 @@ export class GuaranteeListComponent extends IvtListComponent<Guarantee> {
   checked = false;
   indeterminate = false;
   setOfCheckedId = new Set<number>();
+  expandSet = new Set<number>();
+  columns = columns;
+  colorMaps = colorMaps;
+  iconMaps = iconMaps;
+  guaranteeStatus = GuaranteeStatus;
+  statusLabels = statusLabels;
   @Output() downloadPdf = new EventEmitter<number>();
   @Output() createPaymentOrder = new EventEmitter<number[]>();
   @Output() filterCompanies = new EventEmitter<number[]>();
@@ -33,14 +47,28 @@ export class GuaranteeListComponent extends IvtListComponent<Guarantee> {
     return Array.from(this.setOfCheckedId);
   }
 
+  ngOnChanges(changes: SimpleChanges) {
+    if (changes.clearSelected) {
+      this.setOfCheckedId.clear();
+      this.indeterminate = false;
+      this.checked = false;
+    }
+  }
+
+  updateStatusFilter(status: GuaranteeStatus): void {
+    this.filterItems.emit({ status });
+  }
+  onChangeStatus(id: number, status: GuaranteeStatus): void {
+    this.edit.emit({ id, status: GuaranteeStatus[status] });
+  }
+
   onItemChecked(id: number, checked: boolean): void {
     this.updateCheckedSet(id, checked);
     this.refreshCheckedStatus();
-
   }
 
   onAllChecked(value: boolean): void {
-    this.list.forEach(item => this.updateCheckedSet(item.id, value));
+    this.list.filter(item => !item.amount).forEach(item => this.updateCheckedSet(item.id, value));
     this.refreshCheckedStatus();
   }
 
@@ -57,10 +85,11 @@ export class GuaranteeListComponent extends IvtListComponent<Guarantee> {
     }
   }
 
-  updateStatusFilter(status: GuaranteeStatus): void {
-    this.filterItems.emit({ status });
-  }
-  onChangeStatus(id: number, status: GuaranteeStatus): void {
-    this.edit.emit({ id, status: GuaranteeStatus[status] });
+  onExpandChange(id: number, checked: boolean): void {
+    if (checked) {
+      this.expandSet.add(id);
+    } else {
+      this.expandSet.delete(id);
+    }
   }
 }