+++ libs/ui/u-guarantees/src/lib/components/guarantee-list/guarantee-list.constants.ts
@@ -1,29 +1,79 @@
+import { ApsColumns } from '@arphase/ui';
 import { GuaranteeStatus, Select } from '@ivt/c-data';
-import { IvtColumns } from '@ivt/u-ui';
+
+export const columns: ApsColumns = [
+  {
+    label: 'Folio',
+    prop: 'guarantee.id',
+    colSizes: {
+      xs: 8,
+      md: 4,
+      lg: 2,
+    },
+  },
+  {
+    label: 'VIN',
+    prop: 'guarantee.vin',
+    colSizes: {
+      md: 6,
+      lg: 4,
+    },
+  },
+  {
+    label: 'Compañía',
+    prop: 'company.businessName',
+    colSizes: {
+      md: 4
+    },
+  },
+  {
+    label: 'Factura',
+    prop: 'guarantee.invoiceNumber',
+    colSizes: {
+      md: 4,
+      lg: 3,
+    },
+  },
+  {
+    label: 'Importe',
+    prop: 'guarantee.amount',
+    colSizes: {
+      lg: 4,
+    },
+  },
+  {
+    label: 'Estatus',
+    prop: 'guarantee.status',
+    colSizes: {
+      xs: 12,
+      md: 4,
+      lg: 5,
+    },
+  },
+];
 
 export const menuOptions = [
   {
-    class: 'bg-success',
+    type: 'success',
     label: 'Pagada',
     value: GuaranteeStatus.paid,
   },
   {
-    class: 'bg-warning',
+    type: 'warning',
     label: 'Pendiente',
     value: GuaranteeStatus.outstanding,
   },
   {
-    class: 'bg-alert',
+    type: 'error',
     label: 'Cancelada',
     value: GuaranteeStatus.cancelled,
   },
   {
-    class: 'bg-info',
+    type: 'info',
     label: 'Caducada',
     value: GuaranteeStatus.expired,
   },
 ];
-
 
 export const statusOptions: Select[] = [
   {
@@ -43,3 +93,16 @@ export const statusOptions: Select[] = [
     value: GuaranteeStatus[GuaranteeStatus.expired],
   },
 ];
+
+export const colorMaps: Record<GuaranteeStatus, string> = {
+  [GuaranteeStatus.paid]: 'success',
+  [GuaranteeStatus.outstanding]: 'warning',
+  [GuaranteeStatus.cancelled]: 'error',
+  [GuaranteeStatus.expired]: 'processing',
+};
+export const iconMaps: Record<GuaranteeStatus, string> = {
+  [GuaranteeStatus.paid]: 'check-circle',
+  [GuaranteeStatus.outstanding]: 'exclamation-circle',
+  [GuaranteeStatus.cancelled]: 'close-circle',
+  [GuaranteeStatus.expired]: 'sync',
+};