+++ libs/ui/u-guarantees/src/lib/components/payment-order-dialog/payment-order-dialog.component.ts
@@ -1,8 +1,8 @@
 import { ChangeDetectionStrategy, Component, Input, OnChanges, SimpleChanges } from '@angular/core';
 import { FormArray, FormBuilder, FormGroup } from '@angular/forms';
-import { ApsValidators } from '@arphase/ui';
+import { ApsColumns, ApsValidators } from '@arphase/ui';
 import { Guarantee, PaymentOrder } from '@ivt/c-data';
-import { IvtColumns, IvtFormComponent } from '@ivt/u-ui';
+import { IvtFormComponent } from '@ivt/u-ui';
 
 @Component({
   selector: 'ivt-payment-order-dialog',
@@ -12,29 +12,23 @@ import { IvtColumns, IvtFormComponent } from '@ivt/u-ui';
 })
 export class PaymentOrderDialogComponent extends IvtFormComponent<PaymentOrder> implements OnChanges {
   @Input() selectedIds: number[];
-  columns: IvtColumns = [
+  columns: ApsColumns = [
     {
       label: 'Folio',
-      prop: 'id',
-      sortable: false,
       colSizes: {
-        xs: '2',
+        xs: 4,
       },
     },
     {
       label: 'Fecha de factura',
-      prop: 'invoiceDate',
-      sortable: false,
       colSizes: {
-        xs: '5',
+        xs: 10,
       },
     },
     {
       label: 'Importe',
-      prop: 'amount',
-      sortable: false,
       colSizes: {
-        xs: '5',
+        xs: 10,
       },
     },
   ];
@@ -43,7 +37,6 @@ export class PaymentOrderDialogComponent extends IvtFormComponent<PaymentOrder> 
     super();
     this.form = this.fb.group({
       id: null,
-      distributor: [null, ApsValidators.required],
       guarantees: this.fb.array([]),
     });
   }
@@ -57,7 +50,7 @@ export class PaymentOrderDialogComponent extends IvtFormComponent<PaymentOrder> 
       this.selectedIds.forEach(id => this.guaranteesFormArray.push(this.createGuaranteeForm({ id })));
     }
 
-    if (changes.item && this.item) {
+    if (changes.item && this.item?.guarantees) {
       this.form.patchValue(this.item);
       this.item.guarantees.forEach(guarantee => this.guaranteesFormArray.push(this.createGuaranteeForm(guarantee)));
     }