+++ libs/ui/u-guarantees/src/lib/guarantees.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { ApsAutoErrorModule } from '@arphase/ui';
+import { ApsAutoErrorModule, ApsAutoSelectModule } from '@arphase/ui';
 import { PermissionsModule, VehiclesDataModule } from '@ivt/u-state';
 import {
   IvtAddressFormModule,
@@ -16,6 +16,7 @@ import {
   IvtVirtualScrollModule,
 } from '@ivt/u-ui';
 import { VehicleFormModule } from '@ivt/u-vehicles';
+import { NzAlertModule } from 'ng-zorro-antd/alert';
 import { NzButtonModule } from 'ng-zorro-antd/button';
 import { NzCardModule } from 'ng-zorro-antd/card';
 import { NzCheckboxModule } from 'ng-zorro-antd/checkbox';
@@ -35,6 +36,7 @@ import { NzRadioModule } from 'ng-zorro-antd/radio';
 import { NzSelectModule } from 'ng-zorro-antd/select';
 import { NzSpaceModule } from 'ng-zorro-antd/space';
 import { NzTableModule } from 'ng-zorro-antd/table';
+import { NzTagModule } from 'ng-zorro-antd/tag';
 import { NzToolTipModule } from 'ng-zorro-antd/tooltip';
 import { NzTypographyModule } from 'ng-zorro-antd/typography';
 import { NgxMaskModule } from 'ngx-mask';
@@ -92,6 +94,9 @@ import { GuaranteesComponent } from './guarantees.component';
     NzTableModule,
     NzSpaceModule,
     NzDividerModule,
+    ApsAutoSelectModule,
+    NzAlertModule,
+    NzTagModule,
   ],
   declarations: [
     GuaranteeFormComponent,