+++ libs/ui/u-spa/src/lib/spa.module.ts
@@ -2,11 +2,15 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { NzAvatarModule } from 'ng-zorro-antd/avatar';
 import { NzButtonModule } from 'ng-zorro-antd/button';
+import { NzDividerModule } from 'ng-zorro-antd/divider';
 import { NzDropDownModule } from 'ng-zorro-antd/dropdown';
+import { NzGridModule } from 'ng-zorro-antd/grid';
 import { NzIconModule } from 'ng-zorro-antd/icon';
 import { NzLayoutModule } from 'ng-zorro-antd/layout';
 import { NzMenuModule } from 'ng-zorro-antd/menu';
+import { NzSpaceModule } from 'ng-zorro-antd/space';
 import { NzToolTipModule } from 'ng-zorro-antd/tooltip';
+import { NzTypographyModule } from 'ng-zorro-antd/typography';
 
 import { ProfileMenuComponent } from './components/profile-menu';
 import { SpaRoutingModule } from './spa-routing.module';
@@ -23,6 +27,10 @@ import { SpaComponent } from './spa.component';
     NzLayoutModule,
     NzAvatarModule,
     NzDropDownModule,
+    NzDividerModule,
+    NzSpaceModule,
+    NzTypographyModule,
+    NzGridModule,
   ],
   declarations: [SpaComponent, ProfileMenuComponent],
 })