+++ libs/ui/u-vehicles/src/lib/vehicles.module.ts
@@ -18,7 +18,9 @@ import { NzGridModule } from 'ng-zorro-antd/grid';
 import { NzIconModule } from 'ng-zorro-antd/icon';
 import { NzMessageModule } from 'ng-zorro-antd/message';
 import { NzPageHeaderModule } from 'ng-zorro-antd/page-header';
+import { NzSpaceModule } from 'ng-zorro-antd/space';
 import { NzTableModule } from 'ng-zorro-antd/table';
+import { NzTagModule } from 'ng-zorro-antd/tag';
 import { NzToolTipModule } from 'ng-zorro-antd/tooltip';
 import { NzTypographyModule } from 'ng-zorro-antd/typography';
 
@@ -54,6 +56,8 @@ import { VehiclesComponent } from './vehicles.component';
     NzTypographyModule,
     NzDividerModule,
     NzTableModule,
+    NzSpaceModule,
+    NzTagModule,
   ],
   declarations: [
     VehicleListComponent,